Harden candidate support overlap gating
