+++ src/memory_candidate.rs
@@ -12,6 +12,7 @@ mod apply;
 mod parse;
 pub(crate) mod review;
 mod route;
+mod support;
 
 use apply::{
     promote_candidate_to_memory_with_route, update_candidate_after_lifecycle, CandidateApplyOutcome,
@@ -19,6 +20,7 @@ use apply::{
 use parse::{normalize_memory_type, normalize_scope, normalize_topic_key};
 use parse::{parse_defer_reason, parse_memory_candidates};
 pub(super) use route::{route_candidate, CandidateRoute};
+use support::has_conservative_support_token_overlap;
 
 const MEMORY_CANDIDATE_SYSTEM: &str = "\
 Generate durable memory candidates from extracted observations.
@@ -46,21 +48,6 @@ const AUTO_PROMOTE_UNSAFE_MARKERS: &[&str] = &[
     "sk-",
     "token",
 ];
-const MIN_SUPPORT_TOKEN_OVERLAP: usize = 6;
-const MIN_SUPPORT_TOKEN_RATIO: f64 = 0.72;
-const MAX_SUPPORT_TOKEN_WINDOW_EXTRA: usize = 5;
-const SUPPORT_TOKEN_MIN_CHARS: usize = 4;
-#[rustfmt::skip]
-const SUPPORT_RISK_TOKENS: &[&str] = &[
-    "allow", "allowed", "allows", "cannot", "cant", "could", "couldn", "delete", "deleted",
-    "deletes", "deny", "denied", "denies", "didn", "disable", "disabled", "disables", "doesn", "don",
-    "enable", "enabled", "enables", "fail", "failed", "failing", "fails", "hadn", "hasn",
-    "haven", "ignore", "ignored", "ignores", "isn", "may", "might", "never", "no", "not",
-    "prevent", "prevented", "prevents", "reject", "rejected", "rejects", "remove", "removed",
-    "removes", "shouldn", "skip", "skipped", "skips", "succeed", "succeeded", "succeeds",
-    "success", "wasn", "weren", "without", "won", "wouldn",
-];
-
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) enum MemoryCandidateResult {
     EmptyRange,
@@ -599,146 +586,6 @@ fn is_supported_by_source_observation(
         observation_text.contains(&candidate_text)
             || has_conservative_support_token_overlap(&candidate_text, &observation_text)
     })
-}
-
-fn has_conservative_support_token_overlap(candidate_text: &str, observation_text: &str) -> bool {
-    if contains_support_risk_token(candidate_text) || contains_support_risk_token(observation_text)
-    {
-        return false;
-    }
-    let candidate_tokens = support_tokens(candidate_text);
-    if candidate_tokens.len() < MIN_SUPPORT_TOKEN_OVERLAP {
-        return false;
-    }
-    support_token_segments(observation_text)
-        .into_iter()
-        .any(|observation_tokens| {
-            let Some((matched, start, end)) =
-                ordered_support_window(&candidate_tokens, &observation_tokens)
-            else {
-                return false;
-            };
-            let window_len = end.saturating_sub(start) + 1;
-            matched >= MIN_SUPPORT_TOKEN_OVERLAP
-                && (matched as f64 / candidate_tokens.len() as f64) >= MIN_SUPPORT_TOKEN_RATIO
-                && window_len <= candidate_tokens.len() + MAX_SUPPORT_TOKEN_WINDOW_EXTRA
-        })
-}
-
-fn ordered_support_window(
-    candidate_tokens: &[String],
-    observation_tokens: &[String],
-) -> Option<(usize, usize, usize)> {
-    let mut start = None;
-    let mut end = 0;
-    let mut matched = 0;
-    for candidate in candidate_tokens {
-        let search_from = end + usize::from(start.is_some());
-        let position = observation_tokens
-            .iter()
-            .enumerate()
-            .skip(search_from)
-            .find_map(|(index, observation)| (observation == candidate).then_some(index))?;
-        start.get_or_insert(position);
-        end = position;
-        matched += 1;
-    }
-    start.map(|start| (matched, start, end))
-}
-
-fn support_tokens(text: &str) -> Vec<String> {
-    text.split(|ch: char| !ch.is_ascii_alphanumeric())
-        .filter(|token| token.chars().count() >= SUPPORT_TOKEN_MIN_CHARS)
-        .filter(|token| !is_support_stop_token(token))
-        .map(normalize_support_token)
-        .collect()
-}
-
-fn support_token_segments(text: &str) -> Vec<Vec<String>> {
-    let mut segments = Vec::new();
-    let mut current = Vec::new();
-    let mut token = String::new();
-    for ch in text.chars() {
-        if ch.is_ascii_alphanumeric() {
-            token.push(ch.to_ascii_lowercase());
-            continue;
-        }
-        flush_support_segment_token(&mut token, &mut current, &mut segments);
-        if is_support_clause_boundary_char(ch) {
-            finish_support_segment(&mut current, &mut segments);
-        }
-    }
-    flush_support_segment_token(&mut token, &mut current, &mut segments);
-    finish_support_segment(&mut current, &mut segments);
-    segments
-}
-
-fn flush_support_segment_token(
-    token: &mut String,
-    current: &mut Vec<String>,
-    segments: &mut Vec<Vec<String>>,
-) {
-    if token.is_empty() {
-        return;
-    }
-    if is_support_clause_boundary_token(token) {
-        finish_support_segment(current, segments);
-    } else if token.chars().count() >= SUPPORT_TOKEN_MIN_CHARS && !is_support_stop_token(token) {
-        current.push(normalize_support_token(token));
-    }
-    token.clear();
-}
-
-fn finish_support_segment(current: &mut Vec<String>, segments: &mut Vec<Vec<String>>) {
-    if !current.is_empty() {
-        segments.push(std::mem::take(current));
-    }
-}
-
-fn is_support_clause_boundary_char(ch: char) -> bool {
-    matches!(ch, '.' | ';' | ':' | '?' | '!')
-}
-
-fn is_support_clause_boundary_token(token: &str) -> bool {
-    matches!(
-        token,
-        "although" | "and" | "but" | "however" | "though" | "whereas" | "while"
-    )
-}
-
-fn normalize_support_token(token: &str) -> String {
-    if let Some(stem) = token.strip_suffix("ies") {
-        return format!("{stem}y");
-    }
-    if token.len() > 4 && token.ends_with('s') && !token.ends_with("ss") && !token.ends_with("us") {
-        return token[..token.len() - 1].to_string();
-    }
-    token.to_string()
-}
-
-fn contains_support_risk_token(text: &str) -> bool {
-    text.split(|ch: char| !ch.is_ascii_alphanumeric())
-        .any(|token| SUPPORT_RISK_TOKENS.contains(&token))
-}
-
-fn is_support_stop_token(token: &str) -> bool {
-    matches!(
-        token,
-        "about"
-            | "after"
-            | "also"
-            | "from"
-            | "into"
-            | "must"
-            | "only"
-            | "over"
-            | "that"
-            | "their"
-            | "then"
-            | "this"
-            | "uses"
-            | "with"
-    )
 }
 
 fn contains_auto_promote_unsafe_marker(text: &str) -> bool {
